# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Edgebook",
    "description": "College Football Paper-Betting Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cfb/teams": {
      "post": {
        "tags": [
          "college-football"
        ],
        "summary": "Create Team Endpoint",
        "description": "Create a reusable manually-entered college-football team.",
        "operationId": "create_team_endpoint_cfb_teams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games": {
      "post": {
        "tags": [
          "college-football"
        ],
        "summary": "Create Game Endpoint",
        "description": "Create a scheduled game between two existing teams.",
        "operationId": "create_game_endpoint_cfb_games_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "college-football"
        ],
        "summary": "List Games Endpoint",
        "description": "List games newest-first with optional status filter.",
        "operationId": "list_games_endpoint_cfb_games_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameResponse"
                  },
                  "title": "Response List Games Endpoint Cfb Games Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games/{game_id}": {
      "get": {
        "tags": [
          "college-football"
        ],
        "summary": "Get Game Endpoint",
        "description": "Retrieve a game with its manual markets and quotes.",
        "operationId": "get_game_endpoint_cfb_games__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games/{game_id}/markets": {
      "post": {
        "tags": [
          "college-football"
        ],
        "summary": "Create Market Endpoint",
        "description": "Create a draft spread, moneyline, or total market for a game.",
        "operationId": "create_market_endpoint_cfb_games__game_id__markets_post",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/markets/{market_id}/quotes": {
      "post": {
        "tags": [
          "college-football"
        ],
        "summary": "Create Quote Endpoint",
        "description": "Add a manual quote to a draft market.",
        "operationId": "create_quote_endpoint_cfb_markets__market_id__quotes_post",
        "parameters": [
          {
            "name": "market_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Market Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games/{game_id}/quote-comparison": {
      "get": {
        "tags": [
          "college-football"
        ],
        "summary": "Quote Comparison Endpoint",
        "description": "Compare provider observations without selecting a canonical line.",
        "operationId": "quote_comparison_endpoint_cfb_games__game_id__quote_comparison_get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteComparisonResponse"
                  },
                  "title": "Response Quote Comparison Endpoint Cfb Games  Game Id  Quote Comparison Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games/{game_id}/odds-history": {
      "get": {
        "tags": [
          "college-football"
        ],
        "summary": "Odds History Endpoint",
        "description": "Read chronological, bookmaker-specific pregame odds observations.",
        "operationId": "odds_history_endpoint_cfb_games__game_id__odds_history_get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "market_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MarketType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market Type"
            }
          },
          {
            "name": "selection",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MarketSelection"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Selection"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OddsHistoryResponse"
                  },
                  "title": "Response Odds History Endpoint Cfb Games  Game Id  Odds History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games/{game_id}/result": {
      "put": {
        "tags": [
          "college-football"
        ],
        "summary": "Record Game Result Endpoint",
        "description": "Finalize a game and settle its pending simulated bets atomically.",
        "operationId": "record_game_result_endpoint_cfb_games__game_id__result_put",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameResultCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cfb/games/{game_id}/correction": {
      "put": {
        "tags": [
          "college-football"
        ],
        "summary": "Correct Game Result Endpoint",
        "description": "Correct a finalized score and re-settle all affected bets.\n\nCreates offsetting ledger entries to reverse prior payouts, re-settles\nbets with the corrected score, and records an audit-trail entry.",
        "operationId": "correct_game_result_endpoint_cfb_games__game_id__correction_put",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCorrectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameCreate": {
        "properties": {
          "home_team_id": {
            "type": "string",
            "title": "Home Team Id"
          },
          "away_team_id": {
            "type": "string",
            "title": "Away Team Id"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "sport": {
            "$ref": "#/components/schemas/SportType",
            "default": "CFB"
          }
        },
        "type": "object",
        "required": [
          "home_team_id",
          "away_team_id",
          "scheduled_at"
        ],
        "title": "GameCreate",
        "description": "Payload for manually creating a scheduled game."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TeamCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TeamCreate",
        "description": "Payload for a reusable college-football team."
      },
      "TeamResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "TeamResponse",
        "description": "Public representation of a team catalog record."
      },
...
    }
  }
}